+++ src/KskHeptacomAdminOpenAuth.php
@@ -7,10 +7,14 @@ namespace Heptacom\AdminOpenAuth;
 use Doctrine\DBAL\Connection;
 use Shopware\Core\Framework\Plugin;
 use Shopware\Core\Framework\Plugin\Context\UninstallContext;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 final class KskHeptacomAdminOpenAuth extends Plugin
 {
     public const CONFIG_DENY_PASSWORD_LOGIN = 'KskHeptacomAdminOpenAuth.config.denyPasswordLogin';
+
+    public const CONFIG_ENABLE_UNIFIED_REDIRECT_DOMAIN = 'KskHeptacomAdminOpenAuth.config.enableUnifiedRedirectDomain';
+    public const CONFIG_UNIFIED_REDIRECT_DOMAIN = 'KskHeptacomAdminOpenAuth.config.unifiedRedirectDomain';
 
     /**
      * All plugin tables that should be removed on uninstall.
@@ -31,6 +35,13 @@ final class KskHeptacomAdminOpenAuth extends Plugin
     public function executeComposerCommands(): bool
     {
         return true;
+    }
+
+    public function build(ContainerBuilder $container): void
+    {
+        parent::build($container);
+
+        $this->buildDefaultConfig($container);
     }
 
     public function uninstall(UninstallContext $uninstallContext): void