+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Unreleased
 
+**Added**
+
+* Added configuration options `enableUnifiedRedirectDomain` and `unifiedRedirectDomain` to allow using one unified domain for the IdP redirect to the shop. (copied from 8.1.0)
+
+**Changed**
+
+* Changed logs to use the channel `heptacom_admin_open_auth` instead of the default channel
+
 **Fixed**
 
 * Fixed issues when using Shopware with a custom `SHOPWARE_ADMINISTRATION_PATH_NAME` (GitHub Issue #45)